# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Wati Health API
  description: Wati Health Recipe and Medical AI Platform Backend API
  version: 1.0.0
servers:
  - url: http://localhost:5001/api
    description: Local server
  - url: https://api.wati.health/api
    description: Production server
paths:
...
  /recipes:
    get:
      tags:
        - Recipes
      summary: Search and list recipes
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: query
          schema:
            type: string
          description: Search term for recipes
        - in: query
          name: organization_id
          schema:
            type: string
          description: Filter by organization ID (optional)
        - in: query
          name: page
          schema:
            type: integer
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
          description: Items per page
      responses:
        '200':
          description: List of recipes
          content:
            application/json:
              schema:
                type: object
                properties:
                  recipes:
                    type: array
                    items:
                      $ref: '#/components/schemas/Recipe'
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                      pages:
                        type: integer
                      page:
                        type: integer
                      limit:
                        type: integer
...
  /nutri/recipes:
    get:
      tags:
        - Nutrition
      summary: List recipes visible to the health professional
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: number
          schema:
            type: integer
        - in: query
          name: offset
          schema:
            type: integer
      responses:
        '200':
          description: List of recipes
    post:
      tags:
        - Nutrition
      summary: Create an organization-specific recipe
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeUpdate'
      responses:
        '201':
          description: Recipe created
  /nutri/recipes/{id}:
    put:
      tags:
        - Nutrition
      summary: Update organization-specific recipe
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeUpdate'
      responses:
        '200':
          description: Recipe updated
    delete:
      tags:
        - Nutrition
      summary: Delete organization-specific recipe
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Recipe deleted
  /nutri/patients/search:
    get:
      tags:
        - Nutrition
      summary: Search for a global user by email to assign a plan
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: email
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Patient user details
        '404':
          description: Patient not found
  /nutri/patients:
    get:
      tags:
        - Nutrition
      summary: List unique patients with assigned plans
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of patients
  /nutri/plans:
    get:
      tags:
        - Nutrition
      summary: List nutritional plans created by the professional or their organization
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of plans
    post:
      tags:
        - Nutrition
      summary: Create and assign a nutritional plan to a patient
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NutritionalPlanInput'
      responses:
        '201':
          description: Plan created
  /nutri/plans/{id}:
    get:
      tags:
        - Nutrition
      summary: Get details of a nutritional plan
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Plan details
    put:
      tags:
        - Nutrition
      summary: Update a nutritional plan
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NutritionalPlanInput'
      responses:
        '200':
          description: Plan updated
    delete:
      tags:
        - Nutrition
      summary: Delete a nutritional plan
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Plan deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminApiKey:
      type: apiKey
      in: header
      name: X-Admin-Key
  schemas:
...
    Recipe:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title_es:
          type: string
        title_en:
          type: string
        slug:
          type: string
        organization_id:
          type: string
          format: uuid
          nullable: true
        image_url:
          type: string
        tags:
          type: array
          items:
            type: string
        nutrition:
          type: object
    RecipeUpdate:
      type: object
      properties:
        title_es:
          type: string
        title_en:
          type: string
        slug:
          type: string
        prep_time_minutes:
          type: integer
        cook_time_minutes:
          type: integer
        servings:
          type: integer
        difficulty:
          type: string
          enum: [easy, medium, hard]
        status:
          type: string
          enum: [draft, published, archived]
        sibo_risk_level:
          type: string
          enum: [safe, caution, avoid]
...
    NutritionalPlanInput:
      type: object
      required:
        - patient_id
        - title
        - start_date
        - end_date
      properties:
        patient_id:
          type: string
          format: uuid
        title:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        meals:
          type: array
          items:
            type: object
            properties:
              day:
                type: string
              meals:
                type: array
                items:
                  type: object
                  properties:
                    type:
                      type: string
                    recipeId:
                      type: string
                      format: uuid
                    notes:
                      type: string
        notes:
          type: string
...